NLP
Natural language processing (NLP) can be used to detect spam messages. By analyzing the text of a message, NLP systems can identify keywords, phrases, and other patterns that are commonly used in spam messages. NLP systems can also use semantic analysis to identify suspicious messages, as well as analyze the structure of the message to identify anomalies. Additionally, machine learning algorithms can be used to train the system to recognize patterns in messages that indicate the presence of spam.